+++ src/components/common/App/App.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
 /* eslint-disable no-lone-blocks */
 import './App.css';
 
@@ -20,12 +21,22 @@ import ProtectOfRoute from '../ProtectOfRoute';
 import Preloader from '../Preloader/Preloader';
 import mainApi from '../../../helpers/utils/MainApi';
 import useUserData from '../../../сustomHooks/useUserData';
+import moviesApi from '../../../helpers/utils/MoviesApi';
+import { parseMovieData } from '../../../helpers/utils/utils';
 
 function App() {
-  const { setUserDataAndLogin } = useUserData();
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const [isCheckJwt, setIsCheckJwt] = useState(true);
+  console.log('render App');
   const curUser = useCurrentUser();
+  const { setUserDataAndLogin } = useUserData();
+  const [isCheckJwt, setIsCheckJwt] = useState(true);
+
+  const [allFilms, setAllFilms] = useState<any>([]);
+  const [savedFilms, setSavedFilms] = useState<any>([]);
+
+  const moviesProps = {
+    allFilms, savedFilms, setAllFilms, setSavedFilms,
+  };
+
   const AddHeader = React.memo(() => (
     <>
       <Header />
@@ -40,29 +51,82 @@ function App() {
     </>
   ));
 
+  // проверяю токен
   useEffect(() => {
-    if (!curUser?.loggedIn) {
-      mainApi.checkJWT()
-        .then((res) => {
-          setUserDataAndLogin({ values: res, curUser });
-        })
-        .catch((err) => {
-          console.log(`Не удалось авторизоваться автоматически: ${err}`);
-        })
-        .finally(() => setIsCheckJwt(false));
+    const getDataFilms = async () => {
+      try {
+        const filmsData = await moviesApi.getMovies();
+        const savedFilmsData = (await mainApi.getAllSavedMovies()).data;
+
+        setSavedFilms(savedFilmsData.map((film: any) => ({
+          ...film, btnType: 'movies-card__btn_delete',
+        })));
+
+        setAllFilms(filmsData.map((film: any) => {
+          const parsedFilm = parseMovieData(film);
+          const indexInSaved = savedFilmsData
+            .findIndex((el: any) => el.movieId === parsedFilm.movieId);
+          return indexInSaved > -1
+            ? { ...savedFilmsData[indexInSaved], btnType: 'movies-card__btn_saved' }
+            : { ...parsedFilm, btnType: 'movies-card__btn_save' };
+        }));
+      } catch (err) {
+        console.log('Ошибка при попытке получить данные о фильмах с серверов');
+        console.log(err);
+      }
+    };
+
+    try {
+      if (!curUser?.loggedIn) {
+        mainApi.checkJWT()
+          .then((res) => {
+            setUserDataAndLogin({ values: res, curUser });
+            getDataFilms();
+          })
+          .catch((err) => {
+            console.log(`Не удалось авторизоваться автоматически: ${err}`);
+          })
+          .finally(() => setIsCheckJwt(false));
+      }
+    } catch (err) {
+      console.log('Ошибка при попытке получить данные предыдущего запроса пользователя и данные с серверов');
+      console.log(err);
     }
   }, []);
 
+  // useEffect(() => {
+  //   console.log(allFilms);
+  // }, [allFilms]);
+
   return (
+    // чтобы избежать ложного срабатывания защиты
     isCheckJwt ? <Preloader /> : (
       <Routes>
-        <Route path='/signin' element={<LoginContainer />} />
-        <Route path='/signup' element={<RegisterContainer />} />
+        <Route path='/signin' element={<ProtectOfRoute Element={LoginContainer} />} />
+        <Route path='/signup' element={<ProtectOfRoute Element={RegisterContainer} />} />
         <Route path='/' element={<AddHeader />}>
-          <Route path='/profile' element={<ProtectOfRoute Element={ProfileContainer} />} />
+          <Route path='/profile' element={<ProtectOfRoute Element={ProfileContainer} onlyLogedIn />} />
           <Route path='/' element={<AddFooter />}>
-            <Route path='/saved-movies' element={<ProtectOfRoute Element={MoviesContainer} />} />
-            <Route path='/movies' element={<ProtectOfRoute Element={MoviesContainer} />} />
+            <Route
+              path='/saved-movies'
+              element={(
+                <ProtectOfRoute
+                  Element={MoviesContainer}
+                  onlyLogedIn
+                  data={moviesProps}
+                />
+              )}
+            />
+            <Route
+              path='/movies'
+              element={(
+                <ProtectOfRoute
+                  Element={MoviesContainer}
+                  onlyLogedIn
+                  data={moviesProps}
+                />
+              )}
+            />
             <Route index element={<MainContainer />} />
           </Route>
         </Route>
