+++ src/components/common/ProtectOfRoute.tsx
@@ -1,11 +1,23 @@
+/* eslint-disable react/jsx-props-no-spreading */
 import { Navigate } from 'react-router-dom';
 import React from 'react';
 import { useCurrentUser } from '../../contexts/CurrentUserContext';
 
-export default function ProtectRoute({ Element }: { Element: React.ComponentType, }) {
+// interface IProtectRoute {
+//   Element: React.ComponentType,
+//   onlyLogedIn?: boolean,
+// }
+
+export default function ProtectRoute(
+  { Element, onlyLogedIn, ...props }: any
+) {
   const curUser = useCurrentUser();
 
   return (
-    curUser!.loggedIn ? <Element /> : <Navigate to='/' replace />
+    (onlyLogedIn ? curUser!.loggedIn : !curUser!.loggedIn) ? <Element {...props} /> : <Navigate to='/' replace />
   );
 }
+
+ProtectRoute.defaultProps = {
+  onlyLogedIn: false,
+};
