+++ src/components/mainComponents/Promo/Promo.css
