+++ src/components/mainComponents/Techs/Techs.css
