+++ src/components/others/Logo/Logo.tsx
@@ -5,10 +5,12 @@ import { Link } from 'react-router-dom';
 
 import logo from '../../../images/logo.svg';
 
-export default function Logo() {
+function Logo() {
   return (
     <Link to='/' className='logo'>
       <img src={logo} alt='Логотип' className='logo__img btn-hover active-btn-effect' />
     </Link>
   );
 }
+
+export default Logo;
