+++ src/components/others/MoviesCardList/MoviesCardList.css
