+++ src/components/pages/PageWithFilms/PageWithFilms.tsx
@@ -1,54 +1,100 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import './PageWithFilms.css';
 
-import React, { FormEvent } from 'react';
+import React, { FormEvent, useEffect, useState } from 'react';
 
-import SearchForm from '../../sections/SearchForm/SearchForm';
-import MoviesCardList from '../../sections/MoviesCardList/MoviesCardList';
+import { useLocation } from 'react-router-dom';
+import SearchForm from '../../others/SearchForm/SearchForm';
+import MoviesCardList from '../../others/MoviesCardList/MoviesCardList';
 import Preloader from '../../common/Preloader/Preloader';
+import { getSideEffect } from '../../../helpers/utils/utils';
+import useSearcher from '../../../сustomHooks/useSearcher';
+import useSaveCardBtn from '../../../сustomHooks/useSaveCardBtn';
+import useSetterVisibleFilms from '../../../сustomHooks/useSetterVisibleFilms';
 
 interface IPageWithFilms {
-  filters: Record<string, boolean>,
-  setFilters: (
-    ((newValue: Record<string, boolean>) => void) |
-    ((prev: Record<string, boolean>) => { [x: string]: boolean; })
-  )
-  films: any[],
-  onSearch: (e: FormEvent<HTMLFormElement>, value: string) => void,
-  isFilter: boolean,
-  messageForUser: string,
-  onReset: () => void,
-  onClickAddedContent: () => void,
-  cntFilms: number,
-  userQuery: string,
-  onClickSaveBtn: (data: any) => void;
+  // filters: Record<string, boolean>,
+  // setFilters: (
+  //   ((newValue: Record<string, boolean>) => void) |
+  //   ((prev: Record<string, boolean>) => { [x: string]: boolean; })
+  // )
+  // films: any[],
+  // onSearch: (e: FormEvent<HTMLFormElement>, value: string) => void,
+  // isActiveFilters: boolean,
+  // messageForUser: string,
+  // onReset: () => void,
+  // onClickToAddContent: () => void,
+  // cntFilms: number,
+  // userQuery: string,
+  // onClickSaveBtn: (data: any) => void,
 }
 
 export default function PageWithFilms({
-  filters, setFilters, films, onSearch, messageForUser,
-  isFilter, onReset, onClickAddedContent, cntFilms,
-  userQuery, onClickSaveBtn,
-}: IPageWithFilms) {
+  allFilms = [], savedFilms = [], setAllFilms, setSavedFilms,
+}: any) {
+  console.log('>> render PageWithFilms');
+
+  const location = useLocation();
+
+  const [isSearch, setIsSearch] = useState(true);
+  const [isSavedPage, setIsSavedPage] = useState<any>(location.pathname === '/saved-movies');
+
+  const [onClickToAddContent, cntVisibleFilms, onClickToReset] = useSetterVisibleFilms();
+
+  const [
+    onReset, onSearch, visibleFilms, messageForUser,
+    userQuery, isActiveFilters, filters, setFilters, localQuery,
+  ] = useSearcher({
+    allFilms, savedFilms, isSavedPage,
+  });
+
+  const [onClickSaveBtn] = useSaveCardBtn({
+    allFilms, setAllFilms, setSavedFilms, isSavedPage,
+  });
+
+  // хитрая штука для отслеживания перерендеров
+  // позволяет корректно отображать прелоадер и стейт кнопки
+  useEffect(() => {
+    setIsSearch(true);
+
+    if (visibleFilms.length !== 0 // если фильмы нашлись
+      || (visibleFilms.length === 0 && (userQuery || isActiveFilters) // или их нет, но был запрос
+        && ((isSavedPage) || (userQuery !== localQuery)))) { // и при этом...
+      setIsSearch(false);
+    } else { // иначе установить запланированную проверку попозже
+      clearTimeout(getSideEffect);
+      getSideEffect(() => setIsSearch(false), 5000);
+    }
+  }, [visibleFilms]);
+
+  // // для корректной работы страниц с фильмами
+  useEffect(() => {
+    setIsSavedPage(location.pathname === '/saved-movies');
+  }, [location.pathname]);
+
   return (
     <main className='page-with-films'>
       <SearchForm
         filters={filters}
         setFilters={setFilters}
         onSearch={onSearch}
-        onReset={onReset}
+        onReset={() => {
+          onReset();
+          onClickToReset();
+        }}
         userQuery={userQuery}
+        isSearch={isSearch}
       />
-      {films.length > 0 && (
+      {visibleFilms.length > 0 && (
         <MoviesCardList
-          films={films}
-          onClickAddedContent={onClickAddedContent}
-          cntFilms={cntFilms}
+          films={visibleFilms.slice(0, cntVisibleFilms)}
+          onClickToAddContent={onClickToAddContent}
+          cntAllFilms={visibleFilms.length}
           onClickSaveBtn={onClickSaveBtn}
         />
       )}
-      {(isFilter || userQuery) && films.length === 0
-        && <p className='page-with-films__message-for-user'>{messageForUser}</p>}
-      {/* {films.length === 0 && !userQuery && !isFilter && <Preloader />} */}
+      {!isSearch && visibleFilms.length === 0 && <p className='page-with-films__message-for-user'>{messageForUser}</p>}
+      {isSearch && visibleFilms.length === 0 && <Preloader />}
     </main>
   );
 }
