+++ src/helpers/constants.tsx
@@ -27,3 +27,4 @@ export const inputPasswordSettings = {
 
 export const urlMoviesApi = 'https://api.nomoreparties.co';
 export const urlLocalMainApi = 'http://localhost:3002/';
+export const urlMainApi = 'https://api.search-your-movies.nomoredomainsicu.ru/';
