+++ src/сustomHooks/useForm.tsx
@@ -19,7 +19,7 @@ interface IHandleChangeInput {
 //   fetch: (data: IAllString) => Promise<any>;
 // }
 
-export default function useForm({ fetch, toEndFetch }: any) {
+export default function useForm({ fetch, toEndFetch, sbtSucMsg = '' }: any) {
   const curUser = useCurrentUser();
   const [isValidForm, setIsValidForm] = useState(false);
   const [values, setValues] = useState<any>({});
@@ -59,7 +59,7 @@ export default function useForm({ fetch, toEndFetch }: any) {
     fetch(values)
       .then((res: any) => {
         setResData(res);
-        setSbtMsg('Успешно! Подождите ещё немного и я перекину вас к фильмам ;)');
+        setSbtMsg(`Успешно! ${sbtSucMsg}`);
         // объединить свойства, чтобы прокинуть userName, который возвращает login() и хранит в res
         // return чтобы оттянуть .finally и кнопка submit была заблокирована до конца toEndFetch
         return toEndFetch({ values: { ...values, ...res }, curUser });
